python/requirements.txt
pillow
pyyaml
ultralytics
opencv-python
torch
torchvision
tqdm
